services/weaver/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
httpx==0.26.0
neo4j==5.28.1
asyncpg==0.30.0
redis==5.2.1
pyjwt==2.10.1
